+++ ElectronClient/gui/style/theme/dark.js
@@ -1,0 +1,40 @@
+const darkStyle = {
+	backgroundColor: '#1D2024',
+	backgroundColorTransparent: 'rgba(255,255,255,0.9)',
+	oddBackgroundColor: '#dddddd',
+	color: '#dddddd',
+	colorError: 'red',
+	colorWarn: '#9A5B00',
+	colorFaded: '#777777', // For less important text
+	colorBright: '#ffffff', // For important text
+	dividerColor: '#555555',
+	selectedColor: '#333333',
+	urlColor: '#4E87EE',
+
+	backgroundColor2: '#181A1D',
+	depthColor: 'rgb(200, 200, 200, OPACITY)',
+	color2: '#ffffff',
+	selectedColor2: '#013F74',
+	colorError2: '#ff6c6c',
+
+	raisedBackgroundColor: '#474747',
+	raisedColor: '#ffffff',
+
+	warningBackgroundColor: '#CC6600',
+
+	htmlColor: 'rgb(220,220,220)',
+	htmlBackgroundColor: 'rgb(29,32,36)',
+	htmlDividerColor: '#3D444E',
+	htmlCodeColor: '#ffffff',
+	htmlLinkColor: 'rgb(166,166,255)',
+	htmlTableBackgroundColor: 'rgb(40, 41, 42)',
+	htmlCodeBackgroundColor: 'rgb(47, 48, 49)',
+	htmlCodeBorderColor: 'rgb(70, 70, 70)',
+
+	editorTheme: 'twilight',
+	codeThemeCss: 'atom-one-dark-reasonable.css',
+
+	highlightedColor: '#0066C7',
+};
+
+module.exports = darkStyle;
