+++ ElectronClient/gui/style/theme/dracula.js
@@ -1,0 +1,38 @@
+const draculaStyle = {
+	backgroundColor: '#282a36',
+	backgroundColorTransparent: 'rgba(40, 42, 54, 0.9)',
+	oddBackgroundColor: '#282a36',
+	color: '#f8f8f2', // For regular text
+	colorError: '#ff5555',
+	colorWarn: '#ffb86c',
+	colorFaded: '#6272a4', // For less important text;
+	colorBright: '#50fa7b', // For important text;
+	dividerColor: '#bd93f9',
+	selectedColor: '#44475a',
+	urlColor: '#8be9fd',
+
+	backgroundColor2: '#21222C',
+	depthColor: 'rgb(200, 200, 200, OPACITY)',
+	color2: '#bd93f9',
+	selectedColor2: '#44475a',
+	colorError2: '#ff5555',
+
+	raisedBackgroundColor: '#44475a',
+	raisedColor: '#bd93f9',
+
+	warningBackgroundColor: '#ffb86c',
+
+	htmlColor: '#f8f8f2',
+	htmlBackgroundColor: '#282a36',
+	htmlDividerColor: '#f8f8f2',
+	htmlLinkColor: '#8be9fd',
+	htmlTableBackgroundColor: '#6272a4',
+	htmlCodeBackgroundColor: '#44475a',
+	htmlCodeBorderColor: '#f8f8f2',
+	htmlCodeColor: '#50fa7b',
+
+	editorTheme: 'dracula',
+	codeThemeCss: 'atom-one-dark-reasonable.css',
+};
+
+module.exports = draculaStyle;
