+++ ElectronClient/gui/style/theme/light.js
@@ -1,0 +1,39 @@
+// This is the default theme in Joplin
+const lightStyle = {
+	backgroundColor: '#ffffff',
+	backgroundColorTransparent: 'rgba(255,255,255,0.9)',
+	oddBackgroundColor: '#dddddd',
+	color: '#222222', // For regular text
+	colorError: 'red',
+	colorWarn: '#9A5B00',
+	colorFaded: '#777777', // For less important text
+	colorBright: '#000000', // For important text
+	dividerColor: '#dddddd',
+	selectedColor: '#e5e5e5',
+	urlColor: '#155BDA',
+
+	backgroundColor2: '#162B3D',
+	depthColor: 'rgb(100, 182, 253, OPACITY)',
+	color2: '#f5f5f5',
+	selectedColor2: '#0269C2',
+	colorError2: '#ff6c6c',
+
+	raisedBackgroundColor: '#e5e5e5',
+	raisedColor: '#222222',
+
+	warningBackgroundColor: '#FFD08D',
+
+	htmlColor: '#222222',
+	htmlBackgroundColor: 'white',
+	htmlDividerColor: 'rgb(230,230,230)',
+	htmlLinkColor: 'rgb(80,130,190)',
+	htmlTableBackgroundColor: 'rgb(247, 247, 247)',
+	htmlCodeBackgroundColor: 'rgb(243, 243, 243)',
+	htmlCodeBorderColor: 'rgb(220, 220, 220)',
+	htmlCodeColor: 'rgb(0,0,0)',
+
+	editorTheme: 'chrome',
+	codeThemeCss: 'atom-one-light.css',
+};
+
+module.exports = lightStyle;
