+++ ElectronClient/gui/style/theme/solarizedDark.js
@@ -1,0 +1,38 @@
+const solarizedDarkStyle = {
+	backgroundColor: '#002b36',
+	backgroundColorTransparent: 'rgba(0, 43, 54, 0.9)',
+	oddBackgroundColor: '#073642',
+	color: '#93a1a1', // For regular text
+	colorError: '#dc322f',
+	colorWarn: '#cb4b16',
+	colorFaded: '#657b83', // For less important text;
+	colorBright: '#eee8d5', // For important text;
+	dividerColor: '#586e75',
+	selectedColor: '#073642',
+	urlColor: '#268bd2',
+
+	backgroundColor2: '#073642',
+	depthColor: 'rgb(200, 200, 200, OPACITY)',
+	color2: '#eee8d5',
+	selectedColor2: '#6c71c4',
+	colorError2: '#cb4b16',
+
+	raisedBackgroundColor: '#073642',
+	raisedColor: '#839496',
+
+	warningBackgroundColor: '#b5890055',
+
+	htmlColor: '#93a1a1',
+	htmlBackgroundColor: '#002b36',
+	htmlDividerColor: '#073642',
+	htmlLinkColor: '#268bd2',
+	htmlTableBackgroundColor: '#002b36',
+	htmlCodeBackgroundColor: '#002b36',
+	htmlCodeBorderColor: '#696969',
+	htmlCodeColor: '#fdf6e3',
+
+	editorTheme: 'solarized_dark',
+	codeThemeCss: 'atom-one-dark-reasonable.css',
+};
+
+module.exports = solarizedDarkStyle;
