+++ ElectronClient/gui/style/theme/solarizedLight.js
@@ -1,0 +1,38 @@
+const solarizedLightStyle = {
+	backgroundColor: '#fdf6e3',
+	backgroundColorTransparent: 'rgba(253, 246, 227, 0.9)',
+	oddBackgroundColor: '#eee8d5',
+	color: '#657b83', // For regular text
+	colorError: '#dc322f',
+	colorWarn: '#cb4b16',
+	colorFaded: '#839496', // For less important text;
+	colorBright: '#073642', // For important text;
+	dividerColor: '#eee8d5',
+	selectedColor: '#eee8d5',
+	urlColor: '#268bd2',
+
+	backgroundColor2: '#002b36',
+	depthColor: 'rgb(100, 182, 253, OPACITY)',
+	color2: '#eee8d5',
+	selectedColor2: '#6c71c4',
+	colorError2: '#cb4b16',
+
+	raisedBackgroundColor: '#eee8d5',
+	raisedColor: '#073642',
+
+	warningBackgroundColor: '#b5890055',
+
+	htmlColor: '#657b83',
+	htmlBackgroundColor: '#fdf6e3',
+	htmlDividerColor: '#eee8d5',
+	htmlLinkColor: '#268bd2',
+	htmlTableBackgroundColor: '#fdf6e3',
+	htmlCodeBackgroundColor: '#fdf6e3',
+	htmlCodeBorderColor: '#eee8d5',
+	htmlCodeColor: '#002b36',
+
+	editorTheme: 'solarized_light',
+	codeThemeCss: 'atom-one-light.css',
+};
+
+module.exports = solarizedLightStyle;
